+++ ReactCommon/jsc/.clang-tidy
@@ -1,0 +1,6 @@
+---
+Checks: '>
+clang-diagnostic-*,
+'
+InheritParentConfig: true
+...
